Model Refactoring
Author: OlgaKolos
Date: 2017-11-29 16.00
All DB requests are done with Model using.
The check of file deletion is made.
Added deletion of the catalog with the album number.

+++ app/Http/Controllers/EditAlbumController.php
@@ -7,6 +7,7 @@ use App\Classes\PerformerId;
 use App\Http\Requests\EditAlbumRequest;
 use App\M2mPerformerTrack;
 use App\Performer;
+use App\Track;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -46,7 +47,7 @@ class EditAlbumController extends Controller
         $albumYear = $album->getAlbumYear();
         if($request->hasFile('track0')) { //if tracks0 (new track for upload) issets
             $trackKey = 'track0';
-            $tracksKey= DB::table('tracks')->where('album_id', $request->all()['album_id'])->max('track_id') + 1;
+            $tracksKey= Track::where('album_id', $request->all()['album_id'])->max('track_id') + 1;
             $performer = new PerformerId($request->all()['track_performer0']);
             $performerId = $performer->getPerformerId(); //put performer name if needed into DB and get performer id
 
@@ -79,7 +80,7 @@ class EditAlbumController extends Controller
                 }
             }
 
-            DB::table('tracks')->insert([ //put data into tracks table
+            Track::insert([ //put data into tracks table
                 'track_id' => null,
                 'track_name' => $fileName,
                 'track_path' => $tracksKey . "/" .$filePathName,
@@ -87,8 +88,7 @@ class EditAlbumController extends Controller
                 'track_duration' => $trackDuration,
             ]);
 
-            $trackId = DB::table('tracks')-> //get track id
-            where('track_name', $fileName)
+            $trackId = Track::where('track_name', $fileName)//get track id
                 ->where('track_path',$tracksKey . "/" .$filePathName)
                 ->where('album_id', $albumId)
                 ->where('track_duration', $trackDuration)
@@ -105,12 +105,14 @@ class EditAlbumController extends Controller
         foreach ($requestKeys as $requestKey) {
             if(preg_match('/checkbox(\d+)/', $requestKey, $matches)) {//here I get checkboxes for tracks to delete
                 $deleteTrackId = $matches[1];
-                $trackPath = DB::table('tracks') ->where('track_id', $deleteTrackId)->pluck('track_path')[0];
-                DB::table('tracks')->where('track_id', $deleteTrackId)->delete();
-                unlink(public_path() . '/tracks/' . Auth::id() . "/" . $albumId . '/' . $trackPath);//delete track from server
+                $trackPath = Track::where('track_id', $deleteTrackId)->pluck('track_path')[0];
+                Track::where('track_id', $deleteTrackId)->delete();
+                if(file_exists(public_path() . '/tracks/' . Auth::id() . "/" . $albumId . '/' . $trackPath)) {
+                    unlink(public_path() . '/tracks/' . Auth::id() . "/" . $albumId . '/' . $trackPath);//delete track from server
+                }
             }
             else if(preg_match('/track_name(\d+)/', $requestKey, $matches)) {//here I edit tracks name
-                DB::table('tracks')->where('track_id', $matches[1])->
+                Track::where('track_id', $matches[1])->
                     update([
                        "track_name" => $request->all()[$requestKey]
                 ]);
@@ -118,14 +120,14 @@ class EditAlbumController extends Controller
             else if(preg_match('/track_performer(\d+)/', $requestKey, $matches)) {//here I edit tracks performer
                 $performerCorrect = new PerformerId($request->all()[$requestKey]);
                 $idPerformer = $performerCorrect->getPerformerId();
-                DB::table('m2m_performer_tracks')->where('track_id', $matches[1])->
+                M2mPerformerTrack::where('track_id', $matches[1])->
                 update([
                     "performer_id" => $idPerformer
                 ]);
             }
         }
 
-        DB::table('albums')->where('album_id', $albumId)->//here I edit album name and album year
+        Album::where('album_id', $albumId)->//here I edit album name and album year
             update([
                 'album_name' => $albumName,
                 'album_year' => $albumYear
@@ -139,6 +141,27 @@ class EditAlbumController extends Controller
     public function delete(Request $request) {
         $albumId = $request->all()['id'];
         Album::where('album_id', $albumId)->delete();
+
+        $albumPath = public_path() . '/tracks/' . Auth::id() . "/" . $albumId;
+        if (is_dir($albumPath)) {
+            $folders = scandir($albumPath);
+            if(count($folders) >= 2) {
+                foreach($folders as $key1 => $folder){
+                    if($key1 >= 2 and is_dir($albumPath . "/" . $folder)) {
+                        $files = scandir($albumPath . "/" . $folder);
+                        if(count($files >= 2)) {
+                            foreach ($files as $key2 => $file) {
+                                if ($key2 >= 2 and file_exists($albumPath . "/" . $folder . "/" . $file)) {
+                                    unlink($albumPath . "/" . $folder . "/" . $file);
+                                }
+                            }
+                        }
+                        rmdir($albumPath . "/" . $folder);
+                    }
+                }
+            }
+            rmdir($albumPath);
+        }
         return redirect()->route('myAlbums');
     }
 }
